refactor extensions into separate files
