+++ transitions/__init__.py
@@ -1,5 +1,3 @@
 from __future__ import absolute_import
 from .core import (State, Transition, Event, EventData, Machine, MachineError,
                    logger)
-
-from .extensions import (LockedMachine, HierarchicalMachine, LockedHierarchicalMachine, NestedState)
