+++ transitions/extensions/__init__.py
@@ -1,0 +1,4 @@
+from .diagrams import MachineGraphSupport
+from .nesting import HierarchicalMachine, NestedState, NestedTransition
+from .locking import LockedMachine
+from .factory import MachineFactory
