+++ transitions/extensions/diagrams.py
@@ -1,5 +1,105 @@
-from .core import Transition
-from .diagrams import AGraph
+from ..core import Transition, Machine
+import abc
+try:
+    import pygraphviz as pgv
+except:
+    pgv = None
+
+
+class Diagram(object):
+
+    def __init__(self, machine):
+        self.machine = machine
+
+    @abc.abstractmethod
+    def get_graph(self):
+        return
+
+
+class AGraph(Diagram):
+
+    machine_attributes = {
+        'directed': True,
+        'strict': False,
+        'rankdir': 'LR',
+        'ratio': '0.3'
+    }
+
+    style_attributes = {
+        'node': {
+            'default': {
+                'shape': 'circle',
+                'height': '1.2',
+                'style': 'filled',
+                'fillcolor': 'white',
+                'color': 'black',
+            },
+            'active': {
+                'color': 'red',
+                'fillcolor': 'darksalmon',
+                'shape': 'doublecircle'
+            },
+            'previous': {
+                'color': 'blue',
+                'fillcolor': 'azure2',
+            }
+        },
+        'edge': {
+            'default': {
+                'color': 'black',
+
+            },
+            'previous': {
+                'color': 'blue',
+
+            }
+        }
+    }
+
+    def _add_nodes(self, states, container, initial_state=None):
+        # For each state, draw a circle
+        for state in states.keys():
+
+            shape = self.style_attributes['state']['default']['shape']
+
+            container.add_node(n=state, shape=shape)
+
+    def _add_edges(self, events, container):
+        # For each event, add an edge
+        for event in events.items():
+            event = event[1]
+            label = str(event.name)
+
+            for transitions in event.transitions.items():
+                src = transitions[0]
+                for t in transitions[1]:
+                    dst = t.dest
+                    container.add_edge(src, dst, label=label)
+
+    def get_graph(self, title=None):
+        """ Generate a DOT graph with pygraphviz, returns an AGraph object
+        Args:
+            title (string): Optional title for the graph.
+        """
+        if not pgv:
+            raise Exception('AGraph diagram requires pygraphviz')
+
+        if title is None:
+            title = self.__class__.__name__
+        elif title is False:
+            title = ''
+
+        fsm_graph = pgv.AGraph(label=title, **self.machine_attributes)
+        fsm_graph.node_attr.update(self.style_attributes['node']['default'])
+
+        # For each state, draw a circle
+        self._add_nodes(self.machine.states, fsm_graph)
+
+        self._add_edges(self.machine.events, fsm_graph)
+
+        setattr(fsm_graph, 'style_attributes', self.style_attributes)
+
+        return fsm_graph
 
 
 class AAGraph(AGraph):
@@ -41,13 +141,25 @@ class AAGraph(AGraph):
                     sub.add_edge(src, dst, label=label)
 
 
-class MachineGraphSupport(object):
+class MachineGraphSupport(Machine):
+    _pickle_blacklist = ['graph']
 
-    def __init__(self, title='State Machine', *args, **kwargs):
+    def __getstate__(self):
+        return {k: v for k, v in self.__dict__.items() if k not in self._pickle_blacklist}
+
+    def __setstate__(self, state):
+        self.__dict__.update(state)
+        self.graph = self.get_graph(title=self.title)
+        self.set_node_style(self.graph.get_node(self.current_state.name), 'active')
+
+    def __init__(self, *args, **kwargs):
+        # remove title from keywords
+        title = kwargs.pop('title', 'State Machine')
         super(MachineGraphSupport, self).__init__(*args, **kwargs)
 
         # Create graph at beginnning
         self.graph = self.get_graph(title=title)
+        self.title = title
 
         # Set initial node as active
         self.set_node_style(self.graph.get_node(self.initial), 'active')
@@ -107,11 +219,13 @@ class TransitionGraphSupport(Transition):
     def _change_state(self, event_data):
 
         # Mark the active node
-        event_data.machine.set_node_state(self.dest, state='active', reset=True)
+        dest = event_data.machine.get_state(self.dest)
+        dest = dest.name if not hasattr(dest, 'children') else dest.get_initial().name
+        event_data.machine.set_node_state(dest, state='active', reset=True)
 
         # Mark the previous node and path used
         if self.source is not None:
             event_data.machine.set_node_state(self.source, state='previous')
-            event_data.machine.set_edge_state(self.source, self.dest, state='previous')
+            event_data.machine.set_edge_state(self.source, dest, state='previous')
 
         super(TransitionGraphSupport, self)._change_state(event_data)
